Renable Zaqar. (fixed by disabling heat convergence)
This patch re-enables Zaqar which turns out was broken
by Heat convergence being enabled. As we aren't running
it in the undercloud yet either this is fine but we need to
look into this.

Also bumped python-tripleoclient a second time to avoid
an issue with the new --local-copy code there. That has
been reverted and things seem to be working again.