Deploy with new heat-all (engine) container
This patch updates doit.sh to include several new patches
for python-tripleoclient that add the capability to launch
heat-all as a container. This saves us from installing
a bunch of packages (50-60 MB I think) and since we are
installing heat anyways makes a good bit of sense.

The option to use a locally installed heat-all binary
is still supported because the installer still supports
a non-docker mode as well. Do note that I've removed
the heat-api and engine packages from this script so
if you do want to run on baremetal you'll need to
manually install those yourself.

The new python-tripleoclient also captures log files
into a directory (printed out at the end) and
displays heat events during the installation.