Copy undercloud.conf to home directory
doit.sh was trying to do some replacements on undercloud.conf,
however, that file was on the repo root, not on the home directory.

This copies the undercloud.conf from the repo to the home directory
and makes the replacement work. Note that the copying only takes
place if the file is not already in the home directory.